+++ php/tests/Application/Command/SemanticLintTest/UnknownGlobalConstants.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace A;
+
+const BAR = 5;
+
+FOO;
+BAR;
+\A\FOO;
+\A\BAR;